rework command line
